+++ ryven/ironflow/canvas_widgets/nodes.py
@@ -6,18 +6,18 @@ from __future__ import annotations
 
 import numpy as np
 
-from ryven.ironflow.canvas.base import CanvasWidget
-from ryven.ironflow.canvas.buttons import (
+from ryven.ironflow.canvas_widgets.base import CanvasWidget
+from ryven.ironflow.canvas_widgets.buttons import (
     RepresentButtonWidget, ExpandButtonWidget, CollapseButtonWidget, ExecButtonWidget
 )
-from ryven.ironflow.canvas.ports import PortWidget
-from ryven.ironflow.canvas.layouts import NodeLayout, DataPortLayout, ExecPortLayout, ButtonLayout
+from ryven.ironflow.canvas_widgets.ports import PortWidget
+from ryven.ironflow.canvas_widgets.layouts import NodeLayout, DataPortLayout, ExecPortLayout, ButtonLayout
 
 from typing import TYPE_CHECKING, Optional
 if TYPE_CHECKING:
-    from ryven.ironflow.canvas.flow import FlowCanvas
+    from ryven.ironflow.canvas_widgets.flow import FlowCanvas
     from ryven.NENV import Node, NodeInputBP, NodeOutputBP
-    from ryven.ironflow.canvas.base import Number
+    from ryven.ironflow.canvas_widgets.base import Number
 
 __author__ = "Joerg Neugebauer, Liam Huber"
 __copyright__ = (
